remove production configuration
The production configuration was giving me some obtuse errors that
didn't seem to be worth further investigation. I've removed it for
now in favor of the development config. if we ever publish this
to a static site we _still_ might want to use the deveopment version
so sourcemaps and the like are configured